+++ packages/react-storage/src/components/StorageBrowser/views/context/index.ts
@@ -1,0 +1,1 @@
+export { ViewsProvider, useViews } from './views';
